Complete compact-key final validation

+++ storage_compatibility.go
@@ -16,7 +16,6 @@ import (
 	"github.com/cockroachdb/pebble/sstable"
 	"github.com/cockroachdb/pebble/sstable/colblk"
 	"github.com/cockroachdb/pebble/vfs"
-	"github.com/go-bond/bond/utils"
 )
 
 const (
@@ -444,7 +443,7 @@ func WriteStorageCompatibility(databaseDir string, compatibility StorageCompatib
 }
 
 type atomicFileWriteHooks struct {
-	beforePublish func(string) error
+	beforeRename  func(string) error
 	syncDirectory func(string) error
 }
 
@@ -453,11 +452,8 @@ func writeFileAtomically(
 ) error {
 	dir := filepath.Dir(path)
 	existing, err := os.ReadFile(path)
-	if err == nil {
-		if slices.Equal(existing, data) {
-			return syncAtomicFileDirectory(dir, hooks)
-		}
-		return fmt.Errorf("%w at %q", utils.ErrFileContentConflict, path)
+	if err == nil && slices.Equal(existing, data) {
+		return syncAtomicFileDirectory(dir, hooks)
 	}
 	if err != nil && !os.IsNotExist(err) {
 		return err
@@ -468,8 +464,11 @@ func writeFileAtomically(
 		return err
 	}
 	temporaryPath := temporary.Name()
+	renamed := false
 	defer func() {
-		_ = os.Remove(temporaryPath)
+		if !renamed {
+			_ = os.Remove(temporaryPath)
+		}
 	}()
 	if err := temporary.Chmod(mode); err != nil {
 		_ = temporary.Close()
@@ -486,30 +485,15 @@ func writeFileAtomically(
 	if err := temporary.Close(); err != nil {
 		return err
 	}
-	if hooks.beforePublish != nil {
-		if err := hooks.beforePublish(temporaryPath); err != nil {
+	if hooks.beforeRename != nil {
+		if err := hooks.beforeRename(temporaryPath); err != nil {
 			return err
 		}
 	}
-	if err := os.Link(temporaryPath, path); err != nil {
-		if !os.IsExist(err) {
-			return err
-		}
-		existing, readErr := os.ReadFile(path)
-		if readErr != nil {
-			return readErr
-		}
-		if !slices.Equal(existing, data) {
-			return fmt.Errorf("%w at %q", utils.ErrFileContentConflict, path)
-		}
-		return syncAtomicFileDirectory(dir, hooks)
-	}
-	if err := syncAtomicFileDirectory(dir, hooks); err != nil {
+	if err := os.Rename(temporaryPath, path); err != nil {
 		return err
 	}
-	if err := os.Remove(temporaryPath); err != nil {
-		return err
-	}
+	renamed = true
 	return syncAtomicFileDirectory(dir, hooks)
 }
